@@ -15,9 +15,9 @@
  */
 package com.example.android.architecture.blueprints.todoapp.data.source.local
 
-import android.support.test.InstrumentationRegistry
-import android.support.test.filters.LargeTest
-import android.support.test.runner.AndroidJUnit4
+import androidx.test.InstrumentationRegistry
+import androidx.test.filters.LargeTest
+import androidx.test.runner.AndroidJUnit4
 import com.example.android.architecture.blueprints.todoapp.data.Task
 import com.example.android.architecture.blueprints.todoapp.data.source.TasksDataSource
 import com.example.android.architecture.blueprints.todoapp.util.any
@@ -34,7 +34,7 @@ import org.junit.runner.RunWith
 import org.mockito.Mockito.mock
 import org.mockito.Mockito.never
 import org.mockito.Mockito.verify
-import android.arch.persistence.room.Room
+import androidx.room.Room
 import com.example.android.architecture.blueprints.todoapp.utils.SingleExecutors
 
 
