@@ -16,7 +16,7 @@
 package com.example.android.architecture.blueprints.todoapp
 
 import android.content.Context
-import android.support.v4.widget.SwipeRefreshLayout
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import android.util.AttributeSet
 import android.view.View
 
@@ -31,7 +31,7 @@ import android.view.View
 class ScrollChildSwipeRefreshLayout @JvmOverloads constructor(
         context: Context,
         attrs: AttributeSet? = null
-) : SwipeRefreshLayout(context, attrs) {
+) : androidx.swiperefreshlayout.widget.SwipeRefreshLayout(context, attrs) {
 
     var scrollUpChild: View? = null
 
