@@ -15,10 +15,10 @@
  */
 package com.example.android.architecture.blueprints.todoapp
 
-import android.arch.lifecycle.LifecycleOwner
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.Observer
-import android.support.annotation.MainThread
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Observer
+import androidx.annotation.MainThread
 import android.util.Log
 
 import java.util.concurrent.atomic.AtomicBoolean
