@@ -16,10 +16,10 @@
 package com.example.android.architecture.blueprints.todoapp.taskdetail
 
 import android.app.Application
-import android.arch.lifecycle.AndroidViewModel
-import android.databinding.ObservableBoolean
-import android.databinding.ObservableField
-import android.support.annotation.StringRes
+import androidx.lifecycle.AndroidViewModel
+import androidx.databinding.ObservableBoolean
+import androidx.databinding.ObservableField
+import androidx.annotation.StringRes
 import com.example.android.architecture.blueprints.todoapp.R
 import com.example.android.architecture.blueprints.todoapp.SingleLiveEvent
 import com.example.android.architecture.blueprints.todoapp.data.Task
