@@ -15,7 +15,7 @@
  */
 package com.example.android.architecture.blueprints.todoapp.tasks
 
-import android.databinding.BindingAdapter
+import androidx.databinding.BindingAdapter
 import android.widget.ListView
 
 import com.example.android.architecture.blueprints.todoapp.data.Task
